Extract CLI environment variables editor
